+++ projects/igniteui-angular/snackbar/src/snackbar/snackbar.component.ts
@@ -156,7 +156,7 @@ export class IgxSnackbarComponent extends IgxNotificationsDirective
             this.textMessage = message;
         }
 
-        this.strategy = this.container ?
+        this.strategy = this.useContainer ?
             new ContainerPositionStrategy(this.positionSettings) :
             new GlobalPositionStrategy(this.positionSettings);
         super.open();